Prevent out of range

+++ app/src/main/java/eu/faircode/netguard/RuleAdapter.java
@@ -432,8 +432,8 @@ public class RuleAdapter extends RecyclerView.Adapter<RuleAdapter.ViewHolder> im
 
         // Traffic statistics
         holder.tvStatistics.setText(context.getString(R.string.msg_mbday,
-                TrafficStats.getUidTxBytes(rule.info.applicationInfo.uid) * 24 * 3600 * 1000L / 1024f / 1024f / SystemClock.elapsedRealtime(),
-                TrafficStats.getUidRxBytes(rule.info.applicationInfo.uid) * 24 * 3600 * 1000L / 1024f / 1024f / SystemClock.elapsedRealtime()
+                (float) TrafficStats.getUidTxBytes(rule.info.applicationInfo.uid) * 24 * 3600 * 1000 / 1024f / 1024f / SystemClock.elapsedRealtime(),
+                (float) TrafficStats.getUidRxBytes(rule.info.applicationInfo.uid) * 24 * 3600 * 1000 / 1024f / 1024f / SystemClock.elapsedRealtime()
         ));
     }
